// backend/src/audit/enums/audit-action.enum.ts
// src/audit/enums/audit-action.enum.ts

export enum AuditAction {
  LOGIN_SUCCESS = 'LOGIN_SUCCESS',
  LOGIN_FAILED = 'LOGIN_FAILED',
  REGISTER_REQUEST = 'REGISTER_REQUEST',
  REGISTER_CREATED = 'REGISTER_CREATED',
  EMAIL_VERIFIED = 'EMAIL_VERIFIED',
  PASSWORD_RESET_REQUEST = 'PASSWORD_RESET_REQUEST',
  PASSWORD_RESET_COMPLETED = 'PASSWORD_RESET_COMPLETED',
  LOGOUT = 'LOGOUT',
  REFRESH = 'REFRESH',
  TOKEN_REVOKED = 'TOKEN_REVOKED',
  EMAIL_SEND_FAILED = 'EMAIL_SEND_FAILED',
}